backend/requirements.txt
flask==2.3.3
flask-cors==4.0.0
python-dotenv==1.0.0
flask-sqlalchemy==3.1.1
flask-login==0.6.3
werkzeug==2.3.7
email-validator==2.1.0.post1
requests==2.31.0
beautifulsoup4==4.12.2
apscheduler==3.10.4
selenium==4.15.2
webdriver-manager==4.0.1 